[VEUE-864] move source files in to store
