+++ app/javascript/components/media_deck/add_webcam_source.tsx
@@ -1,5 +1,5 @@
 import { Component, h, VNode } from "preact";
-import { WebcamSourceStore } from "components/media_deck/webcam_source";
+import { WebcamSourceStore } from "components/media_deck/store/webcam_source";
 import { observer } from "mobx-preact";
 
 type Props = {
