+++ app/javascript/components/media_deck/store/webcam_source.tsx
